+++ spec/support/entities/author.rb
@@ -1,6 +1,0 @@
-class Author
-  include Datamappify::Entity
-
-  attribute :name, String
-  attribute :bio,  String
-end