+++ spec/support/entities/reversed/author.rb
@@ -1,0 +1,8 @@
+module Reversed
+  class Author
+    include Datamappify::Entity
+
+    attribute :name, String
+    attribute :bio,  String
+  end
+end