+++ spec/support/repositories/active_record/reversed/author_repository.rb
@@ -1,0 +1,8 @@
+module Reversed
+  class AuthorRepositoryActiveRecord
+    include Datamappify::Repository
+
+    for_entity Author
+    default_provider :ActiveRecord
+  end
+end