+++ spec/support/repositories/sequel/author_repository.rb
@@ -1,6 +1,0 @@
-class AuthorRepositorySequel
-  include Datamappify::Repository
-
-  for_entity Author
-  default_provider :Sequel
-end