+++ spec/support/repositories/sequel/reversed/author_repository.rb
@@ -1,0 +1,8 @@
+module Reversed
+  class AuthorRepositorySequel
+    include Datamappify::Repository
+
+    for_entity Author
+    default_provider :Sequel
+  end
+end